+++ services/user.service.js
@@ -1,9 +1,20 @@
 import * as axios from 'axios';
 import { SERVER_ROOT } from 'react-native-dotenv';
 
+const getUserInfo = async (userId, token) => {
+  return axios.get(`${SERVER_ROOT}/users/${userId}`,
+    {
+      headers: { "Authorization": token },
+    })
+    .then((response) => {
+      return response;
+    })
+    .catch((err) => console.error(err));
+};
+
 // GET Request from db, returns promise
 // response --> location: { coords: { lat, lng }, note }
-export const getUserLocation = (payload) => {
+const getUserLocation = (payload) => {
     const accessToken = {
         headers: { 'x-access-token': 'Bearer ' + payload.token }
     }
@@ -17,7 +28,7 @@ export const getUserLocation = (payload) => {
 
 // PUT Request to db, returns promise
 // payload --> location: { coords: { lat, lng }, note }
-export const updateUserLocation = (payload) => {
+const updateUserLocation = (payload) => {
 
     const config = {
         headers: { 'x-access-token': 'Bearer ' + payload.token }
@@ -28,4 +39,10 @@ export const updateUserLocation = (payload) => {
         .then(response => { return response.data })
         .catch(error => { console.error(error) })
     )
+}
+
+module.exports = {
+    getUserInfo,
+    getUserLocation,
+    updateUserLocation
 }
