--- utils/index.js
@@ -1,9 +1,0 @@
-export * from './_utils/computeDistance'
-
-export * from './_utils/getDeviceLocation'
-
-export * from './_utils/geocodeLocation'
-
-export * from './_utils/userInfo'
-
-export * from './_utils/userLocation'
